Add files via upload
This database contains the results of spectrum availability prediction in indoor buildings using LSTM neural networks. The data includes the features used for the prediction (FEATURE), the correlation between the prediction and the actual measurement (CORR. PREDICC PM.), the accuracy of the spectrum availability (PRECISION DISP PM.), the probability of predicting the spectrum availability (PROB. PREDECIR DISPON PM.) and the relative error of the prediction (ERROR REL PM.).